Calculate hash on write
This uses a 128 bit murmur3 hash of the incoming json (normalised).
The hash value is a hex encoded string of the resulting hashed bytes.

Normalised means encoded to json with the fields in alphanumberical
order, with no superfluous whitespace and no newlines or carriage returns.

Note that this would be fragile against:
* adding new fields (which we'd expect to change the hash anyway)

but is robust against:
* re-ordering of fields in the struct
* being called with the fields in any order in the request
* being called with any mix of whitespace, newlines, etc.

This follows the spec at :
https://github.com/Financial-Times/up-jsonhash-spec

Once tweaked and agreed, it's likely the hashing code will move
somewhere shared.